back-end/back-end-102/lesson-8/pyproject.toml
[project]
name = "lesson-8"
version = "0.1.0"
description = ""
authors = [
    {name = "Ansar",email = "issayevansar1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
